requirements.txt
pandas 
numpy 
scikit-learn 
fastapi 
uvicorn 
openai 
httpx 
pyyaml
faker
google-generativeai
python-dotenv
transformers
torch
openai